chore: rename files according to the codebase style
